New class to list and download AWS files.
resolves #5
resolves #14
resolves #15

Takes listing and downloading functionality out of pyart_plot.py
and moves it to a special class for this purpose called NexradLevel2.

Allows for downloading of files across more than one day (i.e.
across 00z).

Fixes bug where filename on AWS was assumed to be a certain length.

Checks first to see if desired filename is already downloaded and of the
same size as the file on AWS.  If so, does not re-download.